Community api

+++ App/spring-backend/demo/src/main/java/com/example/demo/service/CommunityService.java
@@ -2,12 +2,16 @@ package com.example.demo.service;
 
 import com.example.demo.dto.community.CommunityCreateDto;
 import com.example.demo.dto.community.CommunityUpdateDto;
+import com.example.demo.dto.community.response.CommunityResponseDto;
 import com.example.demo.exception.notfound.CommunityNotFoundException;
+import com.example.demo.mapper.CommunityMapper;
 import com.example.demo.model.Community;
 import com.example.demo.exception.AccessDeniedException;
 import com.example.demo.model.User;
 import com.example.demo.repository.CommunityRepository;
 import com.example.demo.repository.PostRepository;
+import com.example.demo.repository.UserRepository;
+import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -16,6 +20,7 @@ import org.springframework.stereotype.Service;
 import java.time.LocalDate;
 import java.time.Period;
 import java.util.List;
+import java.util.UUID;
 
 @Service
 @RequiredArgsConstructor
@@ -35,8 +40,10 @@ public class CommunityService {
     private final CommunityRepository communityRepository;
     private final PostRepository postRepository;
     private final UserService userService;
+    private final CommunityMapper communityMapper;
+    private final UserRepository userRepository;
 
-    public Community addCommunity(CommunityCreateDto dto){
+    public CommunityResponseDto addCommunity(CommunityCreateDto dto){
 
         User user = userService.getAuthenticatedUser();
 
@@ -44,14 +51,27 @@ public class CommunityService {
 
         Community community = new Community();
         community.setAuthor(user);
-        community.setName(dto.getTitle());
+        community.setName(dto.getName());
         community.setTopic(topic.name());
         community.setDescription(dto.getDescription());
-
-        return communityRepository.save(community);
+        community.setDisplayName(dto.getDisplayName());
+        community.setIconUrl(dto.getIconUrl());
+        community.getMembers().add(user);
+        Community savedCommunity=communityRepository.save(community);
+        return toDto(savedCommunity);
     }
     public List<Community> listAllCommunities(){
         return communityRepository.findAll();
+    }
+
+    public CommunityResponseDto toDto(Community community){
+        CommunityResponseDto dto= communityMapper.toResponseDto(community);
+        int memberCount = community.getMembers() != null ? community.getMembers().size() : 1;
+        int postCount = community.getPosts() != null ? community.getPosts().size() : 0;
+        dto.setMemberCount(memberCount);
+        dto.setPostCount(postCount);
+
+        return dto;
     }
 
     public void deleteCommunityByName(String communityName) {
@@ -60,7 +80,10 @@ public class CommunityService {
         if(!communityToDelete.getAuthor().equals(user)){
             throw new AccessDeniedException("You are not allowed to perform this operation");
         }
-        communityRepository.delete(communityToDelete);
+        if(communityToDelete.getPosts()==null||communityToDelete.getPosts().isEmpty()){
+            communityRepository.delete(communityToDelete);
+        }
+
     }
 
     public Community updateCommunity(String communityName, CommunityUpdateDto updateDto){
@@ -76,15 +99,24 @@ public class CommunityService {
             community.setTopic(topic.name());
         }
 
-        community.setDescription(updateDto.getDescription());
+        if(updateDto.getDescription()!=null) {
+            community.setDescription(updateDto.getDescription());
+        }
+
         community.setTopic(updateDto.getTopic());
-        community.setName(updateDto.getName());
+        if(updateDto.getDisplayName()!=null) {
+            community.setDisplayName(updateDto.getDisplayName());
+        }
+        if(updateDto.getIconUrl()!=null) {
+            community.setIconUrl(updateDto.getIconUrl());
+        }
         return communityRepository.save(community);
     }
 
     public Community findByName(String name) {
         return communityRepository.findByName(name)
                 .orElseThrow(() -> new CommunityNotFoundException("Community not found with name=" + name));
+
     }
 
 
@@ -112,6 +144,28 @@ public class CommunityService {
         }
 
         return result;
+    }
+
+    @Transactional
+    public void joinCommunity(String communityName){
+        Community community = communityRepository.findByName(communityName)
+                .orElseThrow(()-> new RuntimeException("Community Not Found"));
+        User user = userService.getAuthenticatedUser();
+        community.getMembers().add(user);
+        communityRepository.save(community);
+
+    }
+
+    @Transactional
+    public void leaveCommunity(String communityName, UUID userId){
+        Community community = communityRepository.findByName(communityName)
+                .orElseThrow(()-> new RuntimeException("Community Not Found"));
+        User user = userRepository.findById(userId)
+                .orElseThrow(()-> new RuntimeException("User Not Found"));
+
+        community.getMembers().remove(user);
+        communityRepository.save(community);
+
     }
 }
